Created a client and tested Network Movement
Created a build to use a my client.
Set up Cube with a Network Transform to test movement over the network.
Updated the MyHub Script to be a MyNetworkManager script.
Set the MyNetworkManager script to Inherit from NetworkManager.
Did three overrides to log out information about the client and the host
machine